backend/app/models/conversation.py
from pydantic import BaseModel


class ConversationResponse(BaseModel):
    id: str
    title: str
    created_at: str
    updated_at: str


class MessageResponse(BaseModel):
    id: str
    role: str
    type: str
    content: str
    created_at: str
