backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
openai>=1.30.0
aiosqlite>=0.20.0
